surfapp--ml-engine/requirements.txt
pandas
scikit-learn
joblib
requests
arrow
python-dotenv